pico: move the Display Command Set enums to a header
