+++ 32blit-pico/display/dbi.cpp
@@ -2,6 +2,7 @@
 #include <math.h>
 
 #include "display.hpp"
+#include "display_commands.hpp"
 
 #include "hardware/clocks.h"
 #include "hardware/dma.h"
@@ -29,46 +30,11 @@ static bool have_vsync = false;
 static bool backlight_enabled = false;
 static uint32_t last_render = 0;
 
-enum MADCTL : uint8_t {
-  // writing to internal memory
-  ROW_ORDER   = 0b10000000,  // MY / y flip
-  COL_ORDER   = 0b01000000,  // MX / x flip
-  SWAP_XY     = 0b00100000,  // AKA "MV"
-
-  // scanning out from internal memory
-  SCAN_ORDER  = 0b00010000,
-  RGB         = 0b00001000,
-  HORIZ_ORDER = 0b00000100
-};
-
 static const uint8_t rotations[]{
   0,                                                                                // 0
   MADCTL::HORIZ_ORDER | MADCTL::SWAP_XY | MADCTL::COL_ORDER,                        // 90
   MADCTL::HORIZ_ORDER | MADCTL::SCAN_ORDER | MADCTL::COL_ORDER | MADCTL::ROW_ORDER, // 180
   MADCTL::SCAN_ORDER | MADCTL::SWAP_XY | MADCTL::ROW_ORDER                          // 270
-};
-
-// standard commands
-enum MIPIDCS
-{
-  Nop              = 0x00,
-  SoftReset        = 0x01,
-  GetAddressMode   = 0x0B,
-  GetPixelFormat   = 0x0C,
-  EnterSleepMode   = 0x10,
-  ExitSleepMode    = 0x11,
-  ExitInvertMode   = 0x20,
-  EnterInvertMode  = 0x21,
-  DisplayOff       = 0x28,
-  DisplayOn        = 0x29,
-  SetColumnAddress = 0x2A,
-  SetRowAddress    = 0x2B,
-  WriteMemoryStart = 0x2C,
-  SetTearOff       = 0x34,
-  SetTearOn        = 0x35,
-  SetAddressMode   = 0x36,
-  SetPixelFormat   = 0x3A,
-  SetTearScanline  = 0x44,
 };
 
 enum ST7789Reg
